added particle system

+++ Mars/Engine/Renderer/RenderPipelineLayers/RenderPipelines/ImGuiRenderPipeline.cpp
@@ -9,31 +9,31 @@
 #include "imgui_impl_sdl2.h"
 #include "imgui_impl_vulkan.h"
 
-void mrs::ImGuiRenderPipeline::Init() 
+void mrs::ImGuiRenderPipeline::Init()
 {
 	_window_handle = (SDL_Window *)Application::GetInstance().GetWindow()->GetNativeWindow();
 
 	// 1: create descriptor pool for IMGUI
 	//  the size of the pool is very oversize, but it's copied from imgui demo itself.
 	VkDescriptorPoolSize pool_sizes[] =
-		{
-			{VK_DESCRIPTOR_TYPE_SAMPLER, 1000},
-			{VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, 1000},
-			{VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE, 1000},
-			{VK_DESCRIPTOR_TYPE_STORAGE_IMAGE, 1000},
-			{VK_DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER, 1000},
-			{VK_DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER, 1000},
-			{VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER, 1000},
-			{VK_DESCRIPTOR_TYPE_STORAGE_BUFFER, 1000},
-			{VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC, 1000},
-			{VK_DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC, 1000},
-			{VK_DESCRIPTOR_TYPE_INPUT_ATTACHMENT, 1000}};
+	{
+		{VK_DESCRIPTOR_TYPE_SAMPLER, 1000},
+		{VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, 1000},
+		{VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE, 1000},
+		{VK_DESCRIPTOR_TYPE_STORAGE_IMAGE, 1000},
+		{VK_DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER, 1000},
+		{VK_DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER, 1000},
+		{VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER, 1000},
+		{VK_DESCRIPTOR_TYPE_STORAGE_BUFFER, 1000},
+		{VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC, 1000},
+		{VK_DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC, 1000},
+		{VK_DESCRIPTOR_TYPE_INPUT_ATTACHMENT, 1000} };
 
 	VkDescriptorPoolCreateInfo pool_info = {};
 	pool_info.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;
 	pool_info.flags = VK_DESCRIPTOR_POOL_CREATE_FREE_DESCRIPTOR_SET_BIT;
 	pool_info.maxSets = 1000;
-	pool_info.poolSizeCount = std::size(pool_sizes);
+	pool_info.poolSizeCount = static_cast<uint32_t>(std::size(pool_sizes));
 	pool_info.pPoolSizes = pool_sizes;
 
 	VkDescriptorPool imguiPool;
@@ -62,26 +62,26 @@ void mrs::ImGuiRenderPipeline::Init()
 
 	// execute a gpu command to upload imgui font textures
 	_renderer->ImmediateSubmit([&](VkCommandBuffer cmd)
-							  { ImGui_ImplVulkan_CreateFontsTexture(cmd); });
+		{ ImGui_ImplVulkan_CreateFontsTexture(cmd); });
 
 	// clear font textures from cpu data
 	ImGui_ImplVulkan_DestroyFontUploadObjects();
 
 	// add the destroy the imgui created structures
 	_renderer->GetDeletionQueue().Push([=]()
-									  {
-		vkDestroyDescriptorPool(_renderer->GetDevice().device, imguiPool, nullptr);
-		ImGui_ImplVulkan_Shutdown(); });
+		{
+			vkDestroyDescriptorPool(_renderer->GetDevice().device, imguiPool, nullptr);
+			ImGui_ImplVulkan_Shutdown(); });
 }
 
-void mrs::ImGuiRenderPipeline::Begin(VkCommandBuffer cmd, uint32_t current_frame) 
+void mrs::ImGuiRenderPipeline::Begin(VkCommandBuffer cmd, uint32_t current_frame)
 {
 	ImGui_ImplVulkan_NewFrame();
 	ImGui_ImplSDL2_NewFrame(_window_handle);
 	ImGui::NewFrame();
 }
 
-void mrs::ImGuiRenderPipeline::End(VkCommandBuffer cmd) 
+void mrs::ImGuiRenderPipeline::End(VkCommandBuffer cmd)
 {
 	ImGui::Render();
 	ImGui_ImplVulkan_RenderDrawData(ImGui::GetDrawData(), cmd);
